Update pom.xml to include latest version number.
Looks like the RPM has an improper version number, keeping it from being easily updated via yum.  This fixes the problem.

Might it be better to use "${project.build_number}" or something similar to include this automatically?  I am not a java programmer so I'm not sure what variables are available at this point.